% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{MCMC_ADRS_AD}
\alias{MCMC_ADRS_AD}
\title{ARMS Individual}
\usage{
MCMC_ADRS_AD(
  y,
  beta,
  beta_a,
  beta_b,
  z_i,
  zeta_a,
  zeta_b,
  lam_ad_z,
  mu_mat_z,
  Sigma_ad_z,
  N,
  K,
  prop_sigma_beta,
  acc_target_z = 0.234,
  scan_every = 100L,
  Iterations = 1000,
  k_shift = 0,
  eq_option = 0,
  eq_prob = 0.5,
  ad_option = 0,
  distr_option = 1
)
}
\arguments{
\item{y}{N x N adjacency matrix}

\item{beta}{scalar, initial value for the intercept}

\item{beta_a}{scalar, prior mean for the intercept}

\item{beta_b}{scalar, prior standard deviation for the intercept}

\item{z_i, }{N x K matrix, initial value for the latent coordinates}

\item{zeta_a}{scalar, prior mean of the latent coordinates, usually set to 0}

\item{zeta_b}{scalar, prior standard deviation of the latent coordinates, usually set to 0}

\item{lam_ad_z}{N vector, lambda parameters, see Andrieu and Thoms 2008, A tutorial on adaptive MCMC}

\item{mu_mat_z}{N x K matrix, mu parameters, see Andrieu and Thoms 2008}

\item{Sigma_ad_z}{List of N K x K matrices,  see Andrieu and Thoms 2008}

\item{N}{scalar, number of nodes}

\item{K}{scalar, number of latent dimensions}

\item{prop_sigma_beta}{scalar, proposal standard deviation for the intercept beta}

\item{acc_target_z}{scalar, target acceptance rate for the latent coordinates, default 0.234}

\item{scan_every}{integer, how many iteration before the update of selection probabilities}

\item{Iterations}{integer, MCMC number of iterations}

\item{k_shift}{scalar, shift of the decision rule sigmoid function}

\item{eq_option}{binary, if 1 selection probabilities with equal default value}

\item{eq_prob}{scalar, between 0 and 1, if eq_option = 1, this is the default value}

\item{ad_option}{binary, if ad_option 0 we use the adaptive proposal by Andrieu and Thoms 2008 otherwise by Latuszynski, Roberts and Rosenthal 2013}

\item{distr_option}{binary, if 0 it runs the algorithm for a LS with Bernoulli likelihood, otherwise with Poisson likelihood, default 1}
}
\value{
List of saved quantities  beta_ite intercpet chain,zeta_ite list of latent positions chains, P_ite matrix of selection probabilities chains, Acc_ite matrix of acceptance rates
}
\description{
The function runs an MCMC algorithm to estimate Latent Space Models using Adaptive Multiple Random Scan.
The latent coordinates get probabilistically updated individually.
}
