% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{procrustes_cpp}
\alias{procrustes_cpp}
\title{Procrustes Trasnformation}
\usage{
procrustes_cpp(X, Y)
}
\arguments{
\item{X}{is an n x d matrix of reference points}

\item{Y}{is an n x d matrix of points to transform}
}
\value{
is an n x d matrix of rotated points
}
\description{
It applies Procrustes transformation of a set of n points laying in a d-dimensional space w.r.t. a reference set.
}
